+++ app/src/main/java/sa/idc/driversapp/domain/interactors/driverTasks/DriverTasksRepository.kt
@@ -8,11 +8,11 @@ interface DriverTasksRepository {
     fun getTasksList(): Single<List<DriverTask>>
 
     /** @return task with given [taskId], null if such task does not exist */
-    fun getTaskById(taskId: Int): Single<DriverTask?>
+    fun getTaskById(taskId: Long): Single<DriverTask?>
 
     /**@return status of accepting the task using [taskId] */
-    fun acceptTaskById(taskId:Int): Single<DriverTasksInteractor.AcceptanceResult>
+    fun acceptTaskById(taskId:Long): Single<DriverTasksInteractor.AcceptanceResult>
 
     /**@return status of finishing the task using [taskID]*/
-    fun finishTaskById(taskId: Int):Single<DriverTasksInteractor.FinishiingResult>
+    fun finishTaskById(taskId: Long):Single<DriverTasksInteractor.FinishiingResult>
 }
